+++ module/src/main/java/com/example/fundInfo/module/mapper/FundInfoMapper.java
@@ -4,6 +4,7 @@ import com.example.fundInfo.module.entity.FundInfo;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
+import java.math.BigInteger;
 import java.util.List;
 
 
@@ -14,5 +15,5 @@ public interface FundInfoMapper {
     List<FundInfo> findAll();
 
     @Select("select * from fund_info where id = #{id} and is_deleted = 0")
-    FundInfo findInfoById(@Param("id") Integer id);
+    FundInfo findInfoById(@Param("id") BigInteger id);
 }
