+++ module/src/main/java/com/example/fundInfo/module/service/FundInfoService.java
@@ -2,11 +2,9 @@ package com.example.fundInfo.module.service;
 
 import com.example.fundInfo.module.entity.FundInfo;
 import com.example.fundInfo.module.mapper.FundInfoMapper;
-import org.apache.ibatis.annotations.Param;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import javax.annotation.Resource;
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -18,7 +16,7 @@ public class FundInfoService {
         return fundInfoMapper.findAll();
     }
 
-    public FundInfo findInfoById(Integer id) {
+    public FundInfo findInfoById(BigInteger id) {
         return fundInfoMapper.findInfoById(id);
     }
 }
